refactor(form): put it into a new directory instead of commons
