+++ src/app/(auth)/signup/_component/signup-user-credential-form.tsx
@@ -4,7 +4,7 @@ import { yupResolver } from '@hookform/resolvers/yup'
 import { useContext, useEffect, useState } from 'react'
 import { useCheckUserExistsByEmailMutation } from '@/lib/features/auth/authApi'
 import { EyeInvisibleOutlined, EyeTwoTone } from '@ant-design/icons'
-import Form from '@/components/commons/form'
+import Form from '@/components/ui/form'
 import { Alert, Button, Divider } from 'antd'
 import { COLOR } from '@/theme/color'
 import { UserSignUpContext } from '../_context/signup-context'
