+++ src/app/(auth)/signup/_component/signup-user-info-form.tsx
@@ -3,7 +3,7 @@ import { useForm } from 'react-hook-form'
 import { yupResolver } from '@hookform/resolvers/yup'
 import { useContext, useEffect, useState } from 'react'
 import { useSignInMutation } from '@/lib/features/auth/authApi'
-import Form from '@/components/commons/form'
+import Form from '@/components/ui/form'
 import { user_info_schema, user_info_type } from '@/lib/schemas/signin.form.dto'
 import { UserSignUpContext } from '../_context/signup-context'
 import { UserInput } from '@/lib/graphql/graphqlTypes'
