+++ src/app/(protected)/chat/chat-page.tsx
@@ -1,0 +1,9 @@
+'use client'
+import Layout from '@/components/layout/layout'
+import React from 'react'
+
+function ChatPage() {
+    return <Layout></Layout>
+}
+
+export default ChatPage
