+++ src/app/(protected)/chat/page.tsx
@@ -1,0 +1,10 @@
+import type { Metadata } from 'next'
+import ChatPage from './chat-page'
+
+export default function IndexPage() {
+    return <ChatPage />
+}
+
+export const metadata: Metadata = {
+    title: 'ChitChat',
+}
